auto formatting

+++ Sim/src/main/Point.java
@@ -1,14 +1,14 @@
 package main;
 
 public class Point {
-	public float x = 0;
-	public float y = 0;
-	
-	public Point(float x, float y) {
+	public double x = 0;
+	public double y = 0;
+
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
-	
+
 	public void Add(Point point) {
 		this.x += point.x;
 		this.y += point.y;
